#3 - break down page code into components
