+++ client/app/layout.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from 'next';
 import './globals.css'
 import { Inter } from 'next/font/google';
+import Navbar from '../components/Navbar';
+import Footer from '@/components/Footer';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
@@ -15,7 +17,12 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Navbar>
+          {children}
+        </Navbar>
+        <Footer />
+        </body>
     </html>
   );
 }
